+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The Azure MCP Server updates automatically by default whenever a new release com
 
 ### Other Changes
 
+- Broaden search for MAC address [[#759](https://github.com/Azure/azure-mcp/pull/759)]
 - Major repository structure change.
   - Service areas moved from `/src/areas/{Area}` and `/tests/areas/{Area}` into `/areas/{area}/src` and `/areas/{area}/tests`
   - Common code moved into `/core/src` and `/core/tests`